@@ -5,14 +5,14 @@ import (
 )
 
 type Response struct {
-	Code    int         `json:"code"`
-	Message string      `json:"message"`
-	Data    interface{} `json:"data,omitempty"`
+	Code    int    `json:"code"`
+	Message string `json:"message"`
+	Data    any    `json:"data,omitempty"`
 }
 
-func NewSuccessResponse(data interface{}) *Response {
+func NewSuccessResponse(data any) *Response {
 	return &Response{
-		Code:    200,
+		Code:    0,
 		Message: "success",
 		Data:    data,
 	}
@@ -25,14 +25,14 @@ func NewErrorResponse(code int, message string) *Response {
 	}
 }
 
-type BetaEvaluateResponse struct {
+type EvaluateResponse struct {
 	Title        string       `json:"title"`
 	Text         [][]string   `json:"text"`
 	EssayInfo    EssayInfo    `json:"essayInfo"`
 	AIEvaluation AIEvaluation `json:"aiEvaluation"`
 }
 
-func (r *BetaEvaluateResponse) JSONString() string {
+func (r *EvaluateResponse) JSONString() string {
 	data, _ := json.Marshal(r)
 	return string(data)
 }
@@ -61,13 +61,15 @@ type Counting struct {
 }
 
 type AIEvaluation struct {
-	ModelVersion           ModelVersion           `json:"modelVersion"`
-	OverallEvaluation      OverallEvaluation      `json:"overallEvaluation"`
-	FluencyEvaluation      FluencyEvaluation      `json:"fluencyEvaluation"`
-	WordSentenceEvaluation WordSentenceEvaluation `json:"wordSentenceEvaluation"`
-	ExpressionEvaluation   ExpressionEvaluation   `json:"expressionEvaluation"`
-	SuggestionEvaluation   SuggestionEvaluation   `json:"suggestionEvaluation"`
-	ParagraphEvaluations   []ParagraphEvaluation  `json:"paragraphEvaluations"`
+	ModelVersion           ModelVersion           `json:"modelVersion,omitempty"`
+	OverallEvaluation      OverallEvaluation      `json:"overallEvaluation,omitempty"`      // 总评
+	FluencyEvaluation      FluencyEvaluation      `json:"fluencyEvaluation,omitempty"`      // 流畅度评价
+	WordSentenceEvaluation WordSentenceEvaluation `json:"wordSentenceEvaluation,omitempty"` // 好词好句评价
+	ExpressionEvaluation   ExpressionEvaluation   `json:"expressionEvaluation,omitempty"`   // 逻辑表达评价
+	SuggestionEvaluation   SuggestionEvaluation   `json:"suggestionEvaluation,omitempty"`   // 建议
+	ParagraphEvaluations   []ParagraphEvaluation  `json:"paragraphEvaluations,omitempty"`   // 段落点评
+	ScoreEvaluation        ScoreEvaluation        `json:"scoreEvaluations,omitempty"`       // 分数点评
+	PolishingEvaluation    []PolishingEvaluation  `json:"polishingEvaluation,omitempty"`    // 润色点评
 }
 
 type ModelVersion struct {
@@ -85,6 +87,83 @@ type FluencyEvaluation struct {
 	FluencyScore       int    `json:"fluencyScore"`
 }
 
+/*
+	{
+		"wordSentenceEvaluation": {
+			"sentenceEvaluations": [
+				[{
+					"isGoodSentence": false,
+					"label": "",
+					"type": {},
+					"wordEvaluations": [{
+						"ori": "，",
+						"revised": "、",
+						"span": [55, 56],
+						"type": {
+							"level1": "还需努力",
+							"level2": "标点问题"
+						}
+					}, {
+						"ori": "，",
+						"revised": "。",
+						"span": [64, 65],
+						"type": {
+							"level1": "还需努力",
+							"level2": "标点问题"
+						}
+					}]
+				}],
+				[{
+					"isGoodSentence": false,
+					"label": "",
+					"type": {},
+					"wordEvaluations": []
+				}, {
+					"isGoodSentence": true,
+					"label": "排比",
+					"type": {
+						"level1": "作文亮点",
+						"level2": "好句"
+					},
+					"wordEvaluations": [{
+						"span": [7, 11],
+						"type": {
+							"level1": "作文亮点",
+							"level2": "好词"
+						}
+					}]
+				}, {
+					"isGoodSentence": true,
+					"label": "比拟",
+					"type": {
+						"level1": "作文亮点",
+						"level2": "好句"
+					},
+					"wordEvaluations": []
+				}, {
+					"isGoodSentence": false,
+					"label": "",
+					"type": {},
+					"wordEvaluations": []
+				}, {
+					"isGoodSentence": true,
+					"label": "比拟",
+					"type": {
+						"level1": "作文亮点",
+						"level2": "好句"
+					},
+					"wordEvaluations": []
+				}, {
+					"isGoodSentence": false,
+					"label": "",
+					"type": {},
+					"wordEvaluations": []
+				}]
+			],
+			"wordSentenceScore": 0
+		}
+	}
+*/
 type WordSentenceEvaluation struct {
 	SentenceEvaluations [][]SentenceEvaluation `json:"sentenceEvaluations"`
 	WordSentenceScore   int                    `json:"wordSentenceScore"`
@@ -93,8 +172,8 @@ type WordSentenceEvaluation struct {
 type SentenceEvaluation struct {
 	IsGoodSentence  bool              `json:"isGoodSentence"`
 	Label           string            `json:"label"`
-	Type            map[string]string `json:"type"`
-	WordEvaluations []WordEvaluation  `json:"wordEvaluations"`
+	Type            map[string]string `json:"type"`            // 好句类型
+	WordEvaluations []WordEvaluation  `json:"wordEvaluations"` // 好词/还需努力的词
 }
 
 type WordEvaluation struct {
@@ -104,6 +183,12 @@ type WordEvaluation struct {
 	Revised string            `json:"revised,omitempty"`
 }
 
+/*
+	"expressionEvaluation": {
+				"expressDescription": "作文能围绕主题展开，但逻辑表达较松散，缺乏层次感。描述场景时重复信息较多（如"放风筝"），未能有效组织细节。人物活动描写琐碎，未形成连贯叙事。建议学习如何筛选关键细节，构建更有条理的场景描写，避免重复和碎片化表达。",
+				"expressionScore": 2
+			},
+*/
 type ExpressionEvaluation struct {
 	ExpressDescription string `json:"expressDescription"`
 	ExpressionScore    int    `json:"expressionScore"`
@@ -116,6 +201,43 @@ type SuggestionEvaluation struct {
 type ParagraphEvaluation struct {
 	ParagraphIndex int    `json:"paragraphIndex"`
 	Comment        string `json:"comment"`
+}
+
+type ScoreEvaluation struct {
+	Comment  string   `json:"comment"`
+	Comments Comments `json:"comments"`
+	Scores   Scores   `json:"scores"`
+}
+
+type Comments struct {
+	Appearance  string `json:"appearance"`
+	Content     string `json:"content"`
+	Expression  string `json:"expression"`
+	Structure   string `json:"structure,omitempty"`   // 结构-初中
+	Development string `json:"development,omitempty"` // 发展-高中
+}
+
+type Scores struct {
+	All         int `json:"all"`
+	Appearance  int `json:"appearance"`
+	Content     int `json:"content"`
+	Expression  int `json:"expression"`
+	Structure   int `json:"structure,omitempty"`
+	Development int `json:"development,omitempty"`
+}
+
+type PolishingEvaluation struct {
+	ParagraphIndex int             `json:"paragraphIndex"`
+	Edits          []PolishingEdit `json:"edits"`
+}
+
+type PolishingEdit struct {
+	Op            string `json:"op"`
+	Reason        string `json:"reason"`
+	Original      string `json:"original"`
+	Revised       string `json:"revised,omitempty"`
+	SentenceIndex int    `json:"sentenceIndex"`
+	Span          []int  `json:"span"`
 }
 
 type TitleOcrResponse struct {
@@ -137,6 +259,49 @@ type DefaultOcrResponse struct {
 }
 
 func (r *DefaultOcrResponse) JSONString() (string, error) {
+	data, err := json.Marshal(r)
+	if err != nil {
+		return "", err
+	}
+	return string(data), nil
+}
+
+// StreamEvaluateResponse 流式评估响应
+type StreamEvaluateResponse struct {
+	Type      string      `json:"type"`      // 响应类型: "init", "progress", "complete", "error"
+	Step      string      `json:"step"`      // 当前步骤: "essay_info", "overall", "fluency", "word_sentence", "expression", "suggestion", "paragraph", "grammar"
+	Progress  int         `json:"progress"`  // 进度百分比 (0-100)
+	Data      interface{} `json:"data"`      // 具体数据
+	Message   string      `json:"message"`   // 状态消息
+	Timestamp int64       `json:"timestamp"` // 时间戳
+}
+
+// StreamInitData 初始化数据
+type StreamInitData struct {
+	Title     string     `json:"title"`
+	Text      [][]string `json:"text"`
+	EssayInfo EssayInfo  `json:"essay_info"`
+}
+
+// StreamStepData 步骤完成数据
+type StreamStepData struct {
+	Step string      `json:"step"`
+	Data interface{} `json:"data"`
+}
+
+// StreamCompleteData 完成数据
+type StreamCompleteData struct {
+	Result *EvaluateResponse `json:"result"`
+}
+
+// StreamErrorData 错误数据
+type StreamErrorData struct {
+	Error string `json:"error"`
+	Step  string `json:"step"`
+}
+
+// JSONString 序列化流式响应
+func (r *StreamEvaluateResponse) JSONString() (string, error) {
 	data, err := json.Marshal(r)
 	if err != nil {
 		return "", err
