@@ -2,7 +2,7 @@ package model
 
 import (
 	"time"
-	
+
 	"go.mongodb.org/mongo-driver/bson/primitive"
 )
 
@@ -12,4 +12,4 @@ type RawLogs struct {
 	Request    string             `bson:"request"`
 	Response   string             `bson:"response"`
 	CreateTime time.Time          `bson:"create_time"`
-} 
+}
