@@ -2,6 +2,9 @@ package logger
 
 import (
 	"essay-stateless/internal/config"
+	"fmt"
+	"path/filepath"
+	"runtime"
 
 	"github.com/sirupsen/logrus"
 )
@@ -14,9 +17,22 @@ func Init(config config.LogConfig) {
 
 	logrus.SetLevel(level)
 
+	logrus.SetReportCaller(true)
+
 	if config.Format == "json" {
-		logrus.SetFormatter(&logrus.JSONFormatter{})
+		logrus.SetFormatter(&logrus.JSONFormatter{
+			CallerPrettyfier: func(f *runtime.Frame) (string, string) {
+				filename := filepath.Base(f.File)
+				return f.Function, fmt.Sprintf("%s:%d", filename, f.Line)
+			},
+		})
 	} else {
-		logrus.SetFormatter(&logrus.TextFormatter{})
+		logrus.SetFormatter(&logrus.TextFormatter{
+			CallerPrettyfier: func(f *runtime.Frame) (string, string) {
+				filename := filepath.Base(f.File)
+				return f.Function, fmt.Sprintf("%s:%d", filename, f.Line)
+			},
+			FullTimestamp: true,
+		})
 	}
 }
